chage: entity별 자료형 변경

+++ src/main/java/com/limhaekyu/eco_rich_backend/domain/Countries.java
@@ -24,5 +24,5 @@ public class Countries {
 
     // ManyToOne??
     @Column(name = "regions_id")
-    private int regionId;
+    private Long regionId;
 }
